+++ src/core/resource/MeshLoader.hpp
@@ -16,7 +16,7 @@ namespace MeshLoader {
 // Structure to hold data for a single sub-mesh
 struct SubMesh {
    std::vector<Vertex> vertices;
-   std::vector<uint16_t> indices;
+   std::vector<uint32_t> indices;
    std::string name;
 };
 
@@ -27,7 +27,7 @@ struct MeshData {
 
    bool IsEmpty() const;
    size_t GetSubMeshCount() const;
-   std::pair<std::vector<Vertex>, std::vector<uint16_t>> GetCombinedData() const;
+   std::pair<std::vector<Vertex>, std::vector<uint32_t>> GetCombinedData() const;
 };
 
 MeshData LoadMesh(const std::string& filepath);
@@ -36,7 +36,7 @@ namespace Internal {
 void ProcessNode(const aiScene* scene, const aiNode* node, std::vector<SubMesh>& subMeshes);
 SubMesh ProcessMesh(const aiScene* scene, const aiMesh* mesh);
 void ExtractVertexData(const aiMesh* mesh, std::vector<Vertex>& vertices);
-void ExtractIndexData(const aiMesh* mesh, std::vector<uint16_t>& indices);
+void ExtractIndexData(const aiMesh* mesh, std::vector<uint32_t>& indices);
    };
 };
