+++ src/vk/resource/VulkanResourceFactory.cpp
@@ -27,7 +27,7 @@ std::unique_ptr<ITexture> VulkanResourceFactory::CreateRenderTarget(const uint32
    return std::make_unique<VulkanTexture>(*m_device, width, height, format, false, samples);
 }
 
-std::unique_ptr<IMesh> VulkanResourceFactory::CreateMesh(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices) {
+std::unique_ptr<IMesh> VulkanResourceFactory::CreateMesh(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices) {
    return std::make_unique<VulkanMesh>(vertices, indices, *m_device);
 }
